requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
requests==2.31.0
python-dotenv==1.0.1
textblob==0.18.0
pandas==2.0.3
python-multipart==0.0.6
openpyxl==3.1.2
